Tandem HinSAGE/Metapath2Vec approach
HinSAGE only predicts the rating value of a hypothetical edge, it does not predict the likelihood of the existence of that edge.  For that, we need Metapath2Vec again.  Together, we can find a new edge that is both likely and likely to be rated highly.

Still having issues with high degree nodes, though.  May need to reintroduct unskewing.